test try 2

+++ fork/au_aiku_worker.php
@@ -35,7 +35,7 @@ $account = false;
 $count=0;
 
 while ($worker->work()) {
-    if ( $worker->returnCode() == GEARMAN_SUCCESS) {
+    if ($count>100 and  $worker->returnCode() == GEARMAN_SUCCESS) {
         $db = null;
         exec("kill -9 ".getmypid());
         die();
